+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Latest
 
 - Add support for automation level to event data model.
+- Add DriverID to event data model in order to support multiple drivers
+  in one journey.
 
 ## 1.0.3